Reduce method definition (#102)
* move lock

* increase method definition for appium

* change include style

* move keyboard

* move ime actions

* move setting

* move context

* move values

* move file managemenet

* move keyevent

* move image comparison

* move app management

* move app state

* move screen record

* remove redundant class/module names

* move clipboard

* shake

* fic rubocop

* tweak
