+++ .rubocop.yml
@@ -14,6 +14,8 @@ Metrics/CyclomaticComplexity:
   Max: 10
 Metrics/PerceivedComplexity:
   Max: 10
+Metrics/ParameterLists:
+  Max: 6
 Style/Documentation:
   Enabled: false
 Style/CommentedKeyword:
