+++ lib/appium_lib_core/android/device/clipboard.rb
@@ -8,8 +8,8 @@ module Appium
           def self.add_methods
             ::Appium::Core::Device.add_endpoint_method(:get_clipboard) do
               def get_clipboard(content_type: :plaintext)
-                unless ::Appium::Core::Device::Clipboard::CONTENT_TYPE.member?(content_type)
-                  raise "content_type should be #{::Appium::Core::Device::Clipboard::CONTENT_TYPE}"
+                unless ::Appium::Core::Base::Device::Clipboard::CONTENT_TYPE.member?(content_type)
+                  raise "content_type should be #{::Appium::Core::Base::Device::Clipboard::CONTENT_TYPE}"
                 end
 
                 params = { contentType: content_type }
@@ -21,8 +21,8 @@ module Appium
 
             ::Appium::Core::Device.add_endpoint_method(:set_clipboard) do
               def set_clipboard(content:, content_type: :plaintext, label: nil)
-                unless ::Appium::Core::Device::Clipboard::CONTENT_TYPE.member?(content_type)
-                  raise "content_type should be #{::Appium::Core::Device::Clipboard::CONTENT_TYPE}"
+                unless ::Appium::Core::Base::Device::Clipboard::CONTENT_TYPE.member?(content_type)
+                  raise "content_type should be #{::Appium::Core::Base::Device::Clipboard::CONTENT_TYPE}"
                 end
 
                 params = {
