+++ lib/appium_lib_core/android/device/screen.rb
@@ -14,7 +14,7 @@ module Appium
               # rubocop:disable Metrics/ParameterLists
               def start_recording_screen(remote_path: nil, user: nil, pass: nil, method: 'PUT', force_restart: nil,
                                          video_size: nil, time_limit: '180', bit_rate: nil, bug_report: nil)
-                option = ::Appium::Core::Device::ScreenRecord.new(
+                option = ::Appium::Core::Base::Device::ScreenRecord.new(
                   remote_path: remote_path, user: user, pass: pass, method: method, force_restart: force_restart
                 ).upload_option
