+++ lib/appium_lib_core/common/base.rb
@@ -1,3 +1,19 @@
+require_relative 'device/device_lock'
+require_relative 'device/keyboard'
+require_relative 'device/ime_actions'
+require_relative 'device/setting'
+require_relative 'device/context'
+require_relative 'device/value'
+require_relative 'device/file_management'
+require_relative 'device/keyevent'
+require_relative 'device/image_comparison'
+require_relative 'device/app_management'
+require_relative 'device/app_state'
+require_relative 'device/screen_record'
+require_relative 'device/battery_status'
+require_relative 'device/clipboard_content_type'
+require_relative 'device/device'
+
 # The following files have selenium-webdriver related stuff.
 require_relative 'base/driver'
 require_relative 'base/bridge'
