+++ lib/appium_lib_core/common/base/driver.rb
@@ -37,6 +37,188 @@ module Appium
           @bridge.dialect
         end
 
+        # Lock
+        def lock(duration = nil)
+          @bridge.lock(duration)
+        end
+
+        def device_locked?
+          @bridge.device_locked?
+        end
+
+        def unlock
+          @bridge.unlock
+        end
+
+        # Keyboard
+        def hide_keyboard(close_key = nil, strategy = nil)
+          @bridge.hide_keyboard close_key, strategy
+        end
+
+        def is_keyboard_shown # rubocop:disable Naming/PredicateName
+          @bridge.is_keyboard_shown
+        end
+
+        # Setting
+        def get_settings
+          @bridge.get_settings
+        end
+
+        def update_settings(settings)
+          @bridge.update_settings(settings)
+        end
+
+        # IME
+        def ime_activate(ime_name)
+          @bridge.ime_activate(ime_name)
+        end
+
+        def ime_available_engines
+          @bridge.ime_available_engines
+        end
+
+        def ime_active_engine
+          @bridge.ime_active_engine
+        end
+
+        def ime_activated
+          @bridge.ime_activated
+        end
+
+        def ime_deactivate
+          @bridge.ime_deactivate
+        end
+
+        # Context
+        def within_context(context)
+          @bridge.within_context(context)
+        end
+
+        def switch_to_default_context
+          @bridge.switch_to_default_context
+        end
+
+        def current_context
+          @bridge.current_context
+        end
+
+        def available_contexts
+          @bridge.available_contexts
+        end
+
+        def set_context(context = null)
+          @bridge.set_context(context)
+        end
+
+        # Value
+        def set_immediate_value(element, *value)
+          @bridge.set_immediate_value(element, *value)
+        end
+
+        def replace_value(element, *value)
+          @bridge.replace_value(element, *value)
+        end
+
+        # File Management
+        def push_file(path, filedata)
+          @bridge.push_file(path, filedata)
+        end
+
+        def pull_file(path)
+          @bridge.pull_file(path)
+        end
+
+        def pull_folder(path)
+          @bridge.pull_folder(path)
+        end
+
+        # Keyevent
+        def keyevent(key, metastate = nil)
+          @bridge.keyevent(key, metastate)
+        end
+
+        def press_keycode(key, metastate: [], flags: [])
+          @bridge.press_keycode(key, metastate: metastate, flags: flags)
+        end
+
+        def long_press_keycode(key, metastate: [], flags: [])
+          @bridge.long_press_keycode(key, metastate: metastate, flags: flags)
+        end
+
+        # App Management
+        def launch_app
+          @bridge.launch_app
+        end
+
+        def close_app
+          @bridge.close_app
+        end
+
+        def reset
+          @bridge.reset
+        end
+
+        def app_strings(language = nil)
+          @bridge.app_strings(language)
+        end
+
+        def background_app(duration = 0)
+          @bridge.background_app(duration)
+        end
+
+        def install_app(path,
+                        replace: nil,
+                        timeout: nil,
+                        allow_test_packages: nil,
+                        use_sdcard: nil,
+                        grant_permissions: nil)
+          @bridge.install_app(path,
+                              replace: replace,
+                              timeout: timeout,
+                              allow_test_packages: allow_test_packages,
+                              use_sdcard: use_sdcard,
+                              grant_permissions: grant_permissions)
+        end
+
+        def remove_app(id, keep_data: nil, timeout: nil)
+          @bridge.remove_app(id, keep_data: keep_data, timeout: timeout)
+        end
+
+        def app_installed?(app_id)
+          @bridge.app_installed?(app_id)
+        end
+
+        def activate_app(app_id)
+          @bridge.activate_app(app_id)
+        end
+
+        def terminate_app(app_id, timeout: nil)
+          @bridge.terminate_app(app_id, timeout: timeout)
+        end
+
+        # AppState
+        def app_state(app_id)
+          @bridge.app_state(app_id)
+        end
+
+        # Screen Record
+        def stop_recording_screen(remote_path: nil, user: nil, pass: nil, method: 'PUT')
+          @bridge.stop_recording_screen(remote_path: remote_path, user: user, pass: pass, method: method)
+        end
+
+        def stop_and_save_recording_screen(file_path)
+          @bridge.stop_and_save_recording_screen(file_path)
+        end
+
+        # Device Handling
+        def shake
+          @bridge.shake
+        end
+
+        def device_time(format = nil)
+          @bridge.device_time(format)
+        end
+
         # Get the device window's size.
         # @return [Selenium::WebDriver::Dimension]
         #
@@ -187,6 +369,36 @@ module Appium
         end
 
         DEFAULT_MATCH_THRESHOLD = 0.5
+
+        # Image Comparison
+        def match_images_features(first_image:,
+                                  second_image:,
+                                  detector_name: 'ORB',
+                                  match_func: 'BruteForce',
+                                  good_matches_factor: nil,
+                                  visualize: false)
+          @bridge.match_images_features(first_image: first_image,
+                                        second_image: second_image,
+                                        detector_name: detector_name,
+                                        match_func: match_func,
+                                        good_matches_factor: good_matches_factor,
+                                        visualize: visualize)
+        end
+
+        def find_image_occurrence(full_image:, partial_image:, visualize: false, threshold: nil)
+          @bridge.find_image_occurrence(full_image: full_image,
+                                        partial_image: partial_image,
+                                        visualize: visualize,
+                                        threshold: threshold)
+        end
+
+        def get_images_similarity(first_image:, second_image:, visualize: false)
+          @bridge.get_images_similarity(first_image: first_image, second_image: second_image, visualize: visualize)
+        end
+
+        def compare_images(mode: :matchFeatures, first_image:, second_image:, options: nil)
+          @bridge.compare_images(mode: mode, first_image: first_image, second_image: second_image, options: options)
+        end
 
         # Return ImageElement if current view has a partial image
         #
