+++ lib/appium_lib_core/common/device/clipboard_content_type.rb
@@ -1,0 +1,11 @@
+module Appium
+  module Core
+    class Base
+      module Device
+        module Clipboard
+          CONTENT_TYPE = [:plaintext, :image, :url].freeze
+        end
+      end # module Clipboard
+    end # class Base
+  end # module Core
+end # module Appium
