--- lib/appium_lib_core/device/clipboard_content_type.rb
@@ -1,9 +1,0 @@
-module Appium
-  module Core
-    module Device
-      module Clipboard
-        CONTENT_TYPE = [:plaintext, :image, :url].freeze
-      end
-    end
-  end
-end
