+++ lib/appium_lib_core/ios/xcuitest/device/performance.rb
@@ -17,10 +17,9 @@ module Appium
               end
 
               ::Appium::Core::Device.add_endpoint_method(:get_performance_record) do
-                # rubocop:disable Metrics/ParameterLists
                 def get_performance_record(save_file_path: './performance', profile_name: 'Activity Monitor',
                                            remote_path: nil, user: nil, pass: nil, method: 'PUT')
-                  option = ::Appium::Core::Device::ScreenRecord.new(
+                  option = ::Appium::Core::Base::Device::ScreenRecord.new(
                     remote_path: remote_path, user: user, pass: pass, method: method
                   ).upload_option
 
@@ -29,7 +28,6 @@ module Appium
 
                   File.open("#{save_file_path}.zip", 'wb') { |f| f << result.unpack('m')[0] }
                 end
-                # rubocop:enable Metrics/ParameterLists
               end
             end
           end # module Performance
