+++ lib/appium_lib_core/ios/xcuitest/device/screen.rb
@@ -9,7 +9,7 @@ module Appium
                 # rubocop:disable Metrics/ParameterLists
                 def start_recording_screen(remote_path: nil, user: nil, pass: nil, method: nil, force_restart: nil,
                                            video_type: 'mp4', time_limit: '180', video_quality: 'medium')
-                  option = ::Appium::Core::Device::ScreenRecord.new(
+                  option = ::Appium::Core::Base::Device::ScreenRecord.new(
                     remote_path: remote_path, user: user, pass: pass, method: method, force_restart: force_restart
                   ).upload_option
