+++ lib/appium_lib_core/ios_xcuitest.rb
@@ -5,5 +5,3 @@ require_relative 'ios/device'
 require_relative 'ios/xcuitest/search_context'
 require_relative 'ios/xcuitest/device'
 require_relative 'ios/xcuitest/bridge'
-
-require_relative 'device/battery_status'
