+++ test/unit/android/device/mjsonwp/commands_test.rb
@@ -427,18 +427,20 @@ class AppiumLibCoreTest
 
           def test_set_immediate_value
             stub_request(:post, "#{SESSION}/appium/element/id/value")
+              .with(body: { value: ["abc\ue000"] }.to_json)
               .to_return(headers: HEADER, status: 200, body: { value: '' }.to_json)
 
-            @driver.set_immediate_value ::Selenium::WebDriver::Element.new(@driver.send(:bridge), 'id'), 'value'
+            @driver.set_immediate_value ::Selenium::WebDriver::Element.new(@driver.send(:bridge), 'id'), %w(a b c)
 
             assert_requested(:post, "#{SESSION}/appium/element/id/value", times: 1)
           end
 
           def test_replace_value
             stub_request(:post, "#{SESSION}/appium/element/id/replace_value")
+              .with(body: { value: ["abc\ue000"] }.to_json)
               .to_return(headers: HEADER, status: 200, body: { value: '' }.to_json)
 
-            @driver.replace_value ::Selenium::WebDriver::Element.new(@driver.send(:bridge), 'id'), 'value'
+            @driver.replace_value ::Selenium::WebDriver::Element.new(@driver.send(:bridge), 'id'), %w(a b c)
 
             assert_requested(:post, "#{SESSION}/appium/element/id/replace_value", times: 1)
           end
@@ -770,6 +772,56 @@ class AppiumLibCoreTest
             assert_requested(:post, "#{SESSION}/execute", times: 1)
             assert_equal :charging, info[:state]
             assert_equal 1.0, info[:level]
+          end
+
+          # IMEs
+          def test_ime_activate
+            stub_request(:post, "#{SESSION}/ime/activate")
+              .with(body: { engine: 'engine name' }.to_json)
+              .to_return(headers: HEADER, status: 200, body: { value: {} }.to_json)
+
+            @driver.ime_activate 'engine name'
+
+            assert_requested(:post, "#{SESSION}/ime/activate", times: 1)
+          end
+
+          def test_ime_available_engines
+            stub_request(:get, "#{SESSION}/ime/available_engines")
+              .to_return(headers: HEADER, status: 200, body: { value: %w(ime1 ime2) }.to_json)
+
+            imes = @driver.ime_available_engines
+
+            assert_requested(:get, "#{SESSION}/ime/available_engines", times: 1)
+            assert_equal imes[0], 'ime1'
+          end
+
+          def test_ime_active_engine
+            stub_request(:get, "#{SESSION}/ime/active_engine")
+              .to_return(headers: HEADER, status: 200, body: { value: 'ime' }.to_json)
+
+            ime = @driver.ime_active_engine
+
+            assert_requested(:get, "#{SESSION}/ime/active_engine", times: 1)
+            assert_equal ime, 'ime'
+          end
+
+          def test_ime_activated
+            stub_request(:get, "#{SESSION}/ime/activated")
+              .to_return(headers: HEADER, status: 200, body: { value: 'true' }.to_json)
+
+            @driver.ime_activated
+
+            assert_requested(:get, "#{SESSION}/ime/activated", times: 1)
+          end
+
+          def test_ime_deactivate
+            stub_request(:post, "#{SESSION}/ime/deactivate")
+              .with(body: {}.to_json)
+              .to_return(headers: HEADER, status: 200, body: { value: 'true' }.to_json)
+
+            @driver.ime_deactivate
+
+            assert_requested(:post, "#{SESSION}/ime/deactivate", times: 1)
           end
         end # class Commands
       end # module MJSONWP
